+++ app/components/followAccount.module.css
@@ -1,0 +1,32 @@
+.followAccount {
+  border: solid #000;
+  border-width: 0 0 0.5px 0;
+
+  .avatar {
+    width: 3rem;
+    height: 3rem;
+    
+    border: 0.5px solid #666;
+
+    border-radius: 50%;
+  }
+
+  div {
+    display: flex;
+    align-items: center;
+
+    padding: 0.5rem;
+  }
+
+  div:hover {
+    background-color: #f4f4f6;
+  }
+
+  a {
+    width: 100%;
+  }
+
+  .accountNameContainer {
+    margin-left: 0.5rem;
+  }
+}
